+++ packages/design-system/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @elf-framework/design-system
 
+## 0.0.13
+
+### Patch Changes
+
+- update AppLayoutItem resize callback
+
+- Updated dependencies []:
+  - @elf-framework/design-tokens@0.0.13
+
 ## 0.0.11
 
 ### Patch Changes