+++ packages/design-tokens/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @elf-framework/design-tokens
 
+## 0.0.13
+
+### Patch Changes
+
+- update AppLayoutItem resize callback
+
 ## 0.0.11
 
 ### Patch Changes